Separating the PlanetariumDome model into a separate 'domes' application
