+++ domes/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from domes.models import PlanetariumDome
+
+admin.site.register(PlanetariumDome)
