+++ domes/models.py
@@ -1,0 +1,14 @@
+from django.db import models
+
+
+class PlanetariumDome(models.Model):
+    name = models.CharField(max_length=255, unique=True)
+    rows = models.IntegerField()
+    seats_in_row = models.IntegerField()
+
+    @property
+    def capacity(self) -> int:
+        return self.rows * self.seats_in_row
+
+    def __str__(self):
+        return f"{self.name} (capacity: {self.capacity})"
