+++ domes/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from domes.models import PlanetariumDome
+
+
+class PlanetariumDomeSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = PlanetariumDome
+        fields = ("id", "name", "rows", "seats_in_row", "capacity")
