+++ planetarium/admin.py
@@ -2,7 +2,6 @@ from django.contrib import admin
 from planetarium.models import (
     ShowTheme,
     AstronomyShow,
-    PlanetariumDome,
     Reservation,
     ShowSession,
     Ticket,
@@ -10,7 +9,6 @@ from planetarium.models import (
 
 admin.site.register(ShowTheme)
 admin.site.register(AstronomyShow)
-admin.site.register(PlanetariumDome)
 admin.site.register(Reservation)
 admin.site.register(ShowSession)
 admin.site.register(Ticket)
