+++ planetarium/tests/test_models.py
@@ -3,10 +3,10 @@ from unittest import mock
 from django.contrib.auth import get_user_model
 from django.core.exceptions import ValidationError
 from django.test import TestCase
+from domes.models import PlanetariumDome
 from planetarium.models import (
     ShowTheme,
     AstronomyShow,
-    PlanetariumDome,
     Reservation,
     ShowSession,
     Ticket,
@@ -49,26 +49,6 @@ class AstronomyShowModelTests(TestCase):
         self.assertEqual(
             astronomy_show.show_themes.get(name=SHOW_THEME_NAME),
             show_theme
-        )
-
-
-class PlanetariumDomeModelTests(TestCase):
-    def test_str_and_capacity_in_planetarium_dome_model(self):
-        data = {
-            "name": PLANETARIUM_DOME_NAME,
-            "rows": 6,
-            "seats_in_row": 10,
-        }
-        calc_capacity = data["rows"] * data["seats_in_row"]
-        planetarium_dome = PlanetariumDome.objects.create(
-            name=data["name"],
-            rows=data["rows"],
-            seats_in_row=data["seats_in_row"]
-        )
-
-        self.assertEqual(
-            str(planetarium_dome),
-            f"{planetarium_dome.name} (capacity: {calc_capacity})"
         )
