+++ planetarium/tests/test_reservation_api.py
@@ -4,11 +4,12 @@ from django.contrib.auth import get_user_model
 from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APIClient
+
+from domes.models import PlanetariumDome
 from planetarium.models import (
     ShowTheme,
     Reservation,
     ShowSession,
-    PlanetariumDome,
     AstronomyShow
 )
 from planetarium.serializers import ReservationListSerializer
