+++ planetarium/urls.py
@@ -3,7 +3,6 @@ from rest_framework import routers
 from planetarium.views import (
     ShowThemeViewSet,
     AstronomyShowViewSet,
-    PlanetariumDomeViewSet,
     ShowSessionViewSet,
     ReservationViewSet,
 )
@@ -11,7 +10,6 @@ from planetarium.views import (
 router = routers.DefaultRouter()
 router.register("show-themes", ShowThemeViewSet)
 router.register("astronomy-shows", AstronomyShowViewSet)
-router.register("planetarium-domes", PlanetariumDomeViewSet)
 router.register("show-sessions", ShowSessionViewSet)
 router.register("reservations", ReservationViewSet)
