+++ planetarium_service/settings.py
@@ -46,6 +46,7 @@ INSTALLED_APPS = [
     "rest_framework",
     "drf_spectacular",
     "debug_toolbar",
+    "domes",
     "planetarium",
     "user",
 ]
